requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
boto3==1.28.63
python-dotenv==1.0.0
Flask-WTF==1.1.1
Pillow==10.0.1
reportlab==4.0.4
pandas==2.1.1
numpy==1.26.4
ultralytics==8.3.0
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
celery==5.4.0
redis==5.0.4